backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
peewee==3.17.1
psycopg2-binary==2.9.9
pydantic==2.6.1
python-dotenv==1.0.1
